Moving TomlTable. Code coverage increase.
