+++ src/main/java/io/ous/jtoml/Toml.java
@@ -1,7 +1,5 @@
 package io.ous.jtoml;
 
-import io.ous.jtoml.impl.TomlTable;
-
 /**
  * The root TomlTable, a Toml configuration file
  * @author Asafh
